Added Python skeleton project using Poetry
Wow! So, I just discovered the wonder of dealing with dependencies in projects in Python! After some research with pipenv, and conda (which I *do* have experience with), I discovered Poetry after some Googling. I like Poetry the best because it has a nice cargo-like interface even though it is still using virtual environments under the hood. It's not perfect (I had to deal with the underlying project venvs because I accidentally made a 2.7 venv to start), but it is certainly much better!