Added input function to Haskell implementation to mirror Rust implementation
In the process of developing Promptor for Rust, I realized I could implement the input function there in Haskell. It's a small change, but it brings the Haskell implementation back up to parity. Now, I have effectively made a Python-like inpu t function for Haskell as well as Rust :smile: